azureフロントとサーバー接続設定変更
